backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.4.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
gotrue==2.4.2
scikit-learn==1.5.0
shapely==2.0.4
alembic==1.13.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.2
geopandas==0.14.3
